backend/requirements.txt
Flask>=2.3.0,<3.0.0
gunicorn>=20.1.0
flask-cors>=4.0.0
numpy
matplotlib
earthengine-api
requests
python-dotenv
google-generativeai
